# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hearth — family AI workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createProfile
      tags: [profile]
      summary: Create or update profile from free-text description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile created/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProfileInput:
      type: object
      required: [rawText]
      properties:
        rawText:
          type: string
          description: Free-text description of user's role and use cases
    Profile:
      type: object
      required: [id, userId, rawText, role, useCases, tone, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        rawText:
          type: string
        role:
          type: string
        useCases:
          type: array
          items:
            type: string
        tone:
          type: string
        notionConnected:
          type: boolean
        notionAccessToken:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # DASHBOARD
...