# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hearth — family AI workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  userId: string;
  rawText: string;
  role: string;
  useCases: string[];
  tone: string;
  notionConnected?: boolean;
  /** @nullable */
  notionAccessToken?: string | null;
  createdAt: Date;
  updatedAt: Date;
}